actionqueue bugs fixed

+++ ShadeEditor/src/main/java/groop/shade/editor/CasterSite.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.math.Vector2;
 public class CasterSite extends Vector2 {
 
 	public static final String IMAGE_PATH = "assets/caster.png";
-	public static final double DEFAULT_SCALE = 0.5;
+	public static final float DEFAULT_SCALE = 0.5f;
 
 	public double scale;
