+++ ShadeEditor/src/main/java/groop/shade/editor/PlayerSite.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.math.Vector2;
 public class PlayerSite extends Vector2 {
 
 	public static final String IMAGE_PATH = "assets/player.png";
-	public static final double DEFAULT_SCALE = 1.0 / 7.0f;
+	public static final float DEFAULT_SCALE = 1.0f / 7.0f;
 
 	public double scale;
